Fixing #36 -- errors swallowed when coverage turned off (#37)

+++ src/node-jspm-jasmine.js
@@ -366,8 +366,10 @@ function importTestFiles(SystemJS, jasmine, specDir, specFiles, coverage, errCal
 					}
 				})
 				.catch((ex) => {
-					// remove temporary directory
-					rimraf.sync(coverage.tempDirectory);
+					if (coverage && coverage.tempDirectory) {
+						// remove temporary directory
+						rimraf.sync(coverage.tempDirectory);
+					}
 					errCallback(ex);
 				});
 			});
